refactor(extraction): share BASIC subroutine-synthesis helpers
Move the verbatim gwbasic/msbasic2 copies of BasicLine,
find_subroutine_ranges, and derive_function_name into a new
extraction::basic_common module, and fold the shared top-level-line
filtering of extract_top_level_calls into for_each_top_level_line.
No behavior change.
